+++ src/redux/Login/hooks.ts
@@ -8,7 +8,7 @@ import { cleanAfterAuthAction, userLoggedIn } from './slice';
 import { useTypedAsyncFn } from 'redux/useTypedAsyncFn';
 import { getAfterAuthFollowCause, getAfterAuthRedirectTo } from './selectors';
 import { useCallback } from 'react';
-import { useCauseFollow } from 'redux/Cause/hooks';
+import { useCauseFollow } from 'redux/Cause/hooks/useCauseFollow';
 
 export const useAfterAuthAction = () => {
   const causeToFollow = useSelector(getAfterAuthFollowCause);
